services/transcription-service/requirements.txt
# FastAPI stack
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# Transcription
faster-whisper>=1.0.0
ctranslate2>=4.0.0

# Speaker diarization
pyannote.audio>=3.1.0
torch>=2.1.0
torchaudio>=2.1.0

# Speaker name inference (optional)
openai>=1.12.0

# Structured logging
python-json-logger>=3.0.0

# HTTP client
httpx>=0.27.0
